Added the "D"elete key for the Upcoming Recordings, EPG, and all program list pages. If the current item is a match for a record rule, a popup will appear to confirm deletion of that recoring rule.
This is an extention of the idea by cpinkham in [12545] with
a little refinement. The title is taken from the SR object to
show the correct search titles and the RecTypeText() is now
included in the popup messages to help specify the rule that
will be deleted. Also, the signalChange() sends the recordid
so any old recordmatch entries will be cleaned up.


git-svn-id: http://svn.mythtv.org/trunk@12608 4b91aca1-2a1f-0410-a73a-a78f2df169e8