Fixes #1648. Fixes #2209. Fixes #2722. Fixes various Firewire problems.
This is a sync from the mythtv-firewire branch, see it for detailed descriptions.

This removes any existing Firewire cards, these must be recreated with GUID. The Linux firewire recorder used the port and node on the bus the STB was on to identify a recorder, this is not unique across bus resets and is not unique in the presence of firewire hubs. The Mac firewire recorder just randomly selected STB's for the recorders. This changes the code to use the GUID, which is unique to each firewire device, to identify which one to use for each recorder.

We required the user to select the STB model, even when we knew this from the vendor and model ID. This automatically selects the proper STB model if we know it, otherwise we default to a channel changer that will work with any AV/C compliant recorder (though it may be slower).

IEEE-1394 Bus resets were not handled properly, particularly in the Linux firewire recorder. This adds a reset handler which will find the recorder if it is still on the bus and resume recording.

We always leaked memory when using libavc later than 5.3, configure will now detect the version and if MythTV is compiled with 5.3 present it will require it and will not leak memory. MythTV will still compile with older versions, but we will still leak memory if the user later upgrades to 5.3 without updating MythTV.

This also merges the IPTV & Firewire TS packet listener classes to a single class in mpeg/streamlisteners.h and merges some Mac & Linux Firewire classes to avoid code duplication.

There is still no channel scanner for firewire and the automatic bus reset code (to deal with 5C channels) has been disabled in the merge due to remaining problems, use the mythtv-firewire branch if you want the latest and greatest.



git-svn-id: http://svn.mythtv.org/trunk@12618 4b91aca1-2a1f-0410-a73a-a78f2df169e8