Fixes #2944. Adds Program Number only tuning to digital recorders.
Normally we tune using the VCT, SDT or NIT since these are almost always unique accross transports. This change allows you to tune using just the PAT and PMT instead. By default this is disabled, because it is not as reliable. But you can enable it on a per input basis in the input connections editor of mythtv-setup.

I would recommend that this not be enabled when using DVB-S with a rotor, or with any DVB cards with known frequency tuning problems.



git-svn-id: http://svn.mythtv.org/trunk@12619 4b91aca1-2a1f-0410-a73a-a78f2df169e8