Refs #3041
NOTE: This commit may break or negatively impact the performance of mythmusic. I'm committing it in this possibly broken state before it becomes larger and harder to handle.

Includes the following changes:

Change the music_songs lastplay column into a datetime field, it was previously a timestamp which meant it was erroneously updated every time a change was made to the record.

Breaks the scanning code out into it's own class.

Adds a music_directories table to normalise the information currently stored in music_songs filename. The code doesn't take full advantage of this change yet but ultimately it will have some benefits for speed and directory browsing. For the moment, because both new and old code is in use, this change will actually slow down scanning!

Renames all the member variables in metadata.cpp/h to comply with the coding standards and improve readability.



git-svn-id: http://svn.mythtv.org/trunk@12744 4b91aca1-2a1f-0410-a73a-a78f2df169e8