Closes #2811. Implements BBC Series link functionality
With this changeset users of EIT in the UK will now have
useful programid & seriesid information which allows the
scheduler to better detect duplicates of shows, even when
the EPG monkeys manage to describe the same episode 3
different ways.

This adds support for the default_authority that is currently
being broadcast in the SDT, and stores that with the channel info.

Database Update!

This increases the size of seriesid & programid to accomodate
the larger size of the CRID based seriesid & programid info.
Also adds default_authority to the channel table.

UK users wishing to take full advantage of this new functionality
will need to rescan their channels to populate the default
authority information for each channel.




git-svn-id: http://svn.mythtv.org/trunk@12760 4b91aca1-2a1f-0410-a73a-a78f2df169e8