Rudimentary flv playback support for MythWeb.  This currently relies on the "open" but not GPL-compatible player from:
    http://www.jeroenwijering.com/?item=Flash_Video_Player

(Not GPL compatible because it has "no commercial" and "attribution" clauses.)
    
You will have to download it yourself and install it at:

    modules/tv/flvplayer.swf
  
I am currently looking for someone who would be interested in developing a
legally-compatible flv player to include in MythWeb, so please contact me if
you are interested.

To get it to work, just take the mpg or asx download URL and replace the suffix
with flvp (for "flv player") to get a page showing the flv player and your
video.

'''NOTICE:'''  There is currently no checking to clean up child processes, so if
you cancel your download/playback, it WILL leave a copy of ffmpeg running on 
your server until you kill it manually (or presumably it will eventually go away
when it finishes encoding the video, but I don't know what RAM-use ramifications
this might have).


git-svn-id: http://svn.mythtv.org/trunk@12791 4b91aca1-2a1f-0410-a73a-a78f2df169e8