Add a 'schedruntime' column to the JobQueue table.  By default this is filled in with the current time when a new job is inserted into the table via the JobQueue class.  If schedruntime is in the future, the job will stay in the queue until the schedruntime is past.  There isn't any GUI or code in Myth that sets schedruntime in the future currently, this is being added to make it easier for people to schedule jobs in the future if they want to.  If a future job is in the table when the recording is deleted, the pending job is deleted just like any other job for that recording would be deleted.
The mythfrontend status page and the mythbackend web status page have been
updated to display this schedruntime value if a job is set to run in the
future.

Eventually this may be used for something like allowing an auto-transcode job
to fire off X days after the recording is made, so if you watch the recording
soon enough you'll see the original or if you watch it later, you'll get the
copy that was transcoded to save space.



git-svn-id: http://svn.mythtv.org/trunk@12917 4b91aca1-2a1f-0410-a73a-a78f2df169e8