Don't update recorded.filesize inside the main recorder capture loops.  This should fix the issues people are seeing with recording glitches related to scheduler runs blocking updates to the reocrded table.
This also modifies Mainserver::HandleQueryRecordings() to only update call
ProgramInfo::SetFilesize() if the recording's endtime is in the past and the
filesize in the DB is 0, otherwise it just stats the file and gives the
current filesize to the frontend.  This way the frontend (and MythWeb) always
get the current filesize of an in-progress recording even though the filesize
in the DB isn't getting updated during the recording.

One potential side-effect of this is that the AutoExpirer will see these
recordings as 0 bytes until the recording is over, but that shouldn't cause
any issues because the expirer won't try to expire an in-use program anyway.



git-svn-id: http://svn.mythtv.org/trunk@12918 4b91aca1-2a1f-0410-a73a-a78f2df169e8