Statusbox changes by Dan Wilga.
Allow up/down keys to wrap around so hitting up on the top item will wrap
around to the bottom item and hitting down on the bottom will wrap to the top.

Remember the current position in the list when you exit the status screen and
automatically go to that item again when you reenter.

Closes #3089.



git-svn-id: http://svn.mythtv.org/trunk@12920 4b91aca1-2a1f-0410-a73a-a78f2df169e8