Fix category_type for grabbers that use [Ff]ilm in the category field to indicate theatrical motion pictures.
"category_type" probably should have been named program_type
or show_type and should have been an enum with an integer
to indicate series or sports. However, this is not the case.
This field must contain one of these four collections of ASCII
characters: 'movie', 'series', 'sports' or 'tvshow' regardless
of the users native language or local slang for motion pictures.


git-svn-id: http://svn.mythtv.org/trunk@12937 4b91aca1-2a1f-0410-a73a-a78f2df169e8