Fixes #2736. Fixes bug in DataDirect setup of ATSC channels.
datadirect.cpp:update_channel_basic() was sometimes matching channels based on their frequency rather than the ATSC major and minor channels. Unfortunately, the better the DataDirect listings were the worse MythTV would screew up. This simply adds a check to the freqid matching so that it is only used when the ATSC minor channel is '0' which in the ATSC standards means it is actually referring to an NTSC channel.



git-svn-id: http://svn.mythtv.org/trunk@12948 4b91aca1-2a1f-0410-a73a-a78f2df169e8