This adds an option to apply a sync offset when remuxing the video and audio streams together using mplex.
For the moment you have to enable it by setting the variable 'useSyncOffset' 
to 'True' at the top of the mythburn.py script.

I'm not sure what the best way is to calculate exactly what the offset should be
for all files, I currently use the video and audio start times as reported by 
the avcodec library to workout what the sync offset was prior to splitting the
file and use the same offset when remuxing the streams.

I really need somebody who can see the sync problem to give it a try to see if 
I'm on the right track or not. The audio sync difference is just too subtle for
me to see, I guess some people are more sensitive to this than others :-)

Refs #3146.

git-svn-id: http://svn.mythtv.org/trunk@12951 4b91aca1-2a1f-0410-a73a-a78f2df169e8