Add the ability to set the Commercial Detection method on a per-channel basis. This adds a new commmethod field to the channel table and deprecates the use of the commfree field as that has been rolled into commmethod.
Initial patch by 'talktodan@...' in trac, with modifications by me to fix the
channel editor to fit on one screen as well as eliminating the
'Commercial Free' checkbox and adding a 'Commercial Free' option in the new
commmethod combobox.

I also ran a lupdate and lrelease since this adds/modifies some text that
needs to be translated.

Closes #2878.



git-svn-id: http://svn.mythtv.org/trunk@12957 4b91aca1-2a1f-0410-a73a-a78f2df169e8