Don't specify default values for the fromclause/whereclause text fields.
Early MySQL v5.x versions silently allowed this.  Later MySQL versions give
a warning in non-strict mode and error out when the server is in strict mode.

The code that uses this table always specifies both columns so there shouldn't
be a need for the default value if it is going to cause some people to get
errors trying to upgrade their DB.

Closes #3094.

References MySQL docs and bugs:

   http://bugs.mysql.com/bug.php?id=19498
   http://bugs.mysql.com/bug.php?id=26798
   http://dev.mysql.com/doc/refman/5.0/en/blob.html



git-svn-id: http://svn.mythtv.org/trunk@12958 4b91aca1-2a1f-0410-a73a-a78f2df169e8