Added "Custom Priority", an editor to create, test, install and delete specialized priority factors using power priority.
Power priority is similar to power search in that the user can
insert snippets of SQL that the MythTV scheduler will use when
filling in the schedule. However, where power search is used
to create a search rule to find matching shows in the listings,
power priority adds or subtracts priority for any showing from
any rule if it meets the conditions.

The current stand alone options "Close Captioned Priority" and
"Priority When Shown Once" are examples of these kinds of
specialized priority factors and these two options will be
transferred to power priority in a future commit.

Here are some examples of problems that I wanted to solve with
Power priority:

TV Guide Channel is on a high channel where the signal is best
on card 4.

Priority Rule Name: TVGC
Priority Value: 2
program.chanid = 1095 AND capturecard.cardid = 4

VH1 is best on input 1 so I want to give a disadvantage to all
other inputs for any show from any rule on VH1:

Priority Rule Name: VH1
Priority Value: -1
channel.channum = '39' AND cardinput.cardinputid != 1

Conversely, Versus is only bad on 1 and any other card will do:

Priority Rule Name: VS
Priority Value: -2
channel.callsign = 'VS' AND cardinput.cardinputid = 1

Note that the SQL is an expression that evaluates to true or
false. For any show where the condition is true, the Priority
Value is added. These should generally be simple expressions
that focus on specific conditions. While the examples above
each simply match a station or channel with a card or input,
the possibilities are unlimited.


git-svn-id: http://svn.mythtv.org/trunk@13115 4b91aca1-2a1f-0410-a73a-a78f2df169e8