Closes #3180. Adds eit fixup for dvb-c in the Netherlands. Original patch by Peter Martens.
The year and director probably won't be detected correctly,
as per Peter's original patch. I've added what i think is
the correct regexps to the ticket for them.

I'll add a patch shortly that can be with these regexps
so it can be more easily tested by those with the service.

Please reopen the ticket if you have an updated patch that
fixes the detection of year and director.



git-svn-id: http://svn.mythtv.org/trunk@13183 4b91aca1-2a1f-0410-a73a-a78f2df169e8