Add an option to show the album art cover image for a few seconds in the visualizer after each song change.
Also show the album art image on the info banner shown when the visualization 
is shown full screen and 'INFO' is pressed.

There is a new setting on the visualization settings page to turn this 
feature on.

I did things slightly differently than the patch on the ticket. I just
use the banner to show the image rather than switching visualizations.

Closes #2685

git-svn-id: http://svn.mythtv.org/trunk@13235 4b91aca1-2a1f-0410-a73a-a78f2df169e8