Refactor MythMusic's Visualisations. Based on the original patch on ticket #2595 with some changes to make it easier to add the new LibVisual visualisation.
In addition to the changes stated on the ticket it also make it possible for
a visualiser to support more than one effect. Non of the existing visualisers 
do that but it will make it easier to add a new one based on LibVisual which
can have many plugins each providing a new effect/visual.

I also added a new dialog to edit the allowed visualisations list. The order
you place visualisations in the selected list is the order they will be used
in the player unless you have the 'Randomise' feature turned on.

You probably need to do a distclean of mythmusic so some of the changes to
signal/slots get picked up properly by qmake/moc.

Closes #2595.


git-svn-id: http://svn.mythtv.org/trunk@13268 4b91aca1-2a1f-0410-a73a-a78f2df169e8