Add "libvisual" support to MythMusic.
If the option to add libvisual support is enabled in configure it allows
MythMusic to use any visualizer supported by the libvisual library.

Any new visualizers added to your libvisual setup will automatically appear
in MythMusic's list of available visualizers without having to do any
recompiling of MythMusic. You just need to add them to the list of visuals you
want MythMusic to use on the playback settings screen.

This adds libvisual as an optional dependency for MythMusic. You can choose to 
enable or disable libvisual support from the configure script. It's enabled by
default if libvisual and SDL are found by configure. Add --disable-libvisual
if don't want it.

This commit requires that you rerun the mythplugins ./configure script.

Closes #2307.


git-svn-id: http://svn.mythtv.org/trunk@13269 4b91aca1-2a1f-0410-a73a-a78f2df169e8