Fixes #3107
Fixes a segfault in mythgallery when compiled with openGL disabled.

The solution is a bit of a hack - I had three options available:
1) Fix the shortcomings of TriggeredConfigurationGroup
2) Switch to using a standard configuration group if openGL wasn't enabled
3) Create a triggered config group with only one target and no trigger

I went with #3 as it was the simplest solution to this particular problem.

I also fixed another bug spotted by stuarta where the first time the gallery settings where accessed, the opengl options would be displayed even though SlideshowUseOpenGL was false.



git-svn-id: http://svn.mythtv.org/trunk@13277 4b91aca1-2a1f-0410-a73a-a78f2df169e8