Fixes #2391
We weren't linking against the recordedprogram table in cases when the programme start time and recording start time don't match.

This problem dates back to when we started storing the recording start time in the starttime column, as opposed to the programme start time. This was bound to cause some confusion, because starttime in one table may not equal starttime in another.



git-svn-id: http://svn.mythtv.org/trunk@13339 4b91aca1-2a1f-0410-a73a-a78f2df169e8