Allow ripping of multiple tracks to one file.
This adds a new feature to MythMusic that allows you to rip several tracks
to the same file, useful for classical music lovers who want to make sure
all movements of a piece stay together.

You indicate which consecutive tracks you would like to stay together by
highlighting the track in the track list on the rip tracks screen and pressing
SELECT. The track number and length will change to '-' to indicate that the
track will be ripped to the same file as the previous track. You can chain as
many tracks together as you like.

The current implementation doesn't attempt to change the track numbers or
track titles but you can use the metadata editor to manually change them if
you wish.

Wendy not sure if this is exactly what you wanted, there was a few things
in your original patch that I didn't know what they were for. Let me know if
there is anything you would like changed.

Closes #3043.

git-svn-id: http://svn.mythtv.org/trunk@13381 4b91aca1-2a1f-0410-a73a-a78f2df169e8