Fixes #3360. Fixes how we deal with STBs that ignore the power commands completely (i.e. don't send back an unknown command packet).
We used to assume that if we timed out this was a temporary failure and we would keep trying until it succeeded. Now we only try twice, if it fails both times we treat this as if we had gotten an unimplemented command return from the STB.

Tested by David Frascone and Jim Westfall.

git-svn-id: http://svn.mythtv.org/trunk@13424 4b91aca1-2a1f-0410-a73a-a78f2df169e8