Refs #3410. Works around a bug where the VideoOutputXV::DrawUnusedRects() would draw in the wrong place in the EPG.
XFillRectangle is being called with the correct coordinates, and this works when DrawUnusedRects() is the first time in the EPG, but when it is drawn when the EPG is in the wrong part of its event loop it gets offset. DrawUnusedRects() doesn't really do anything unless the aspect ratio or resolution of the video changes, and then it just leaves a little chaff.

git-svn-id: http://svn.mythtv.org/trunk@13435 4b91aca1-2a1f-0410-a73a-a78f2df169e8