Enable plugins in MythBrowser.
Allows MythBrowser to view YouTube video's for example. To do this first
you need to install the Netscape plugins support for Konqueror and install
the flash player.

There is still a problem in that the flash player doesn't show if you click a
link but if you press INFO and enter the URL of the video directly it always
seems to work?.

git-svn-id: http://svn.mythtv.org/trunk@13437 4b91aca1-2a1f-0410-a73a-a78f2df169e8