Fix a problem with plugins not working when a link is clicked in MythBrowser.
Also changes the way visited pages are saved. It just saves the URL now and
not the entire page which was causing problems because although the pages
were hidden they were still being updated and consuming resources etc. It
was a particular problem if the hidden pages had music or videos playing on
them because they would continue to play even when the page wasn't being
shown.

git-svn-id: http://svn.mythtv.org/trunk@13446 4b91aca1-2a1f-0410-a73a-a78f2df169e8