Add the ability to capture MPEG-2 video from the go7007 driver.  The setting to enable this is currently commented out in recordingprofile.cpp.  To turn it on to test, uncomment the following line around line number 825 in libs/libmythtv/recordingprofile.cpp.
    //codecName->addSelection("MPEG-2");

I get lots of "concealing ... MV errors" and "end mismatch" warnings from
libavcodec's decoder when I turn this on even though the video appears to
play fine, so I'm leaving this commented out by default for now.  It is mainly
being added so I can progress further on other related work.  Mplayer plays
the resulting MPEG-2/MP3 .nuv file fine as well but with the same warnings
when I use "mplayer -vc ffmpeg12,mpeg12" to force the libavcodec decoder.

This commit renames the mpeg4scalebitrate setting to scalebitrate so it can
be reused for both MPEG-4 and MPEG-2 encoding.  The old setting is left in
the codecparams table to maintain the ability to rollback to an older version.

References #2972.

One step closer to writing out container formats other than .nuv (like mpeg-ps)



git-svn-id: http://svn.mythtv.org/trunk@13475 4b91aca1-2a1f-0410-a73a-a78f2df169e8