Put in a (temporary) patch in libavcodec to allow decoding packets with multiple MP3 frames in them.  Previously the MP3 decoder's decode_frame() function would indicate that it had consumed all the data in the packet and would cause the MythTV decoder code to ignore the rest of the packet.  Now, decode_frame will indicate how much data it consumed so our decoder can recall the libavcodec routine if necessary.
I am waiting on the guys at ffmpeg-devel for a little guidance in submitting
a patch upstream to them, in case they want to solve the problem differently
than I have.  In order to use another libavcodec program to transcode one of
these 'broken' nuv files, upstream will have to be fixed.  Once upstream is
fixed, then this patch can be reverted.

References #2972.  The fix for the MythTV issue creating these out of spec
packets was fixed in [13510].



git-svn-id: http://svn.mythtv.org/trunk@13511 4b91aca1-2a1f-0410-a73a-a78f2df169e8