Modify NuppelDecoder to use libavcodec to decode the MP3 audio instead of using lame directly.
Also add the ability to decode AC3 audio in the nuppel file if present.  SVN
HEAD can not currently write AC3 to a nuppel file, but that is something else
I have a patch in my tree for.  This is leading to allowing maintaining the
AC3 audio when transcoding from MPEG-2/.mpg to MPEG-4/.nuv.

If you notice any audio issues when playing nuppel files, try reverting this
patch to see if it cures the issue.  I have tested this with transcoded and
'new' nuppel files generated directly from NuppelVideoRecorder and everything
seems to be working correctly.

References #2972.



git-svn-id: http://svn.mythtv.org/trunk@13512 4b91aca1-2a1f-0410-a73a-a78f2df169e8