Fixes #3505. Avoids some OSD resizes on first channel tune with digital tuners.
This is based on a patch by Matt Doran, the only difference is that this snaps the aspect to 4:3 or 16:9 as appropriate to deal with 16:10 and other sligtly off displays and the code was placed in the dummy video initialization so it is only run when it will be useful.

This only works properly when the channel being tuned matches the display aspect ratio and the display has square pixels. But the fallback is to do the same thing we do currently so there is no loss for the many people this doesn't do anything for.


git-svn-id: http://svn.mythtv.org/trunk@13540 4b91aca1-2a1f-0410-a73a-a78f2df169e8