Add an optional idle timer to LiveTV. If we haven't received any remote/keyboard input for X minutes then we show a dialogue asking if the user is still watching. After 45 seconds without any reponse, we exit LiveTV. If the user responds, then the timer is reset.
This is a limited case feature and is designed for users in households where LiveTV is left running even though nobody is watching any more. It differs from the present sleep function because it is always on, it is reset every time a button is pushed and it checks with the user that it is ok to exit first.

There is no "visible" setting for feature as setup menus are already overcrowded. To enable it, insert a row into the settings table for "LiveTVIdleTimeout" with a value representing the number of minutes we should be idle before exiting.



git-svn-id: http://svn.mythtv.org/trunk@13560 4b91aca1-2a1f-0410-a73a-a78f2df169e8